requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
pydantic==1.9
requests==2.31.0
# google
google-auth==2.6.5

# dotenv
python-dotenv==0.20.0

#test
pytest==7.1.0

# db
databases[postgresql]==0.3.1
SQLAlchemy==1.4.27
pg8000==1.24.0

# date
python-dateutil~=2.8.2